project refactoring. [in progress]
