+++ src/App.js
@@ -1,7 +1,7 @@
 import React from "react";
 import "./App.css";
 import Home from "./pages/Home/Home";
-import Profile from "./pages/Home/profile/profile.js";
+import Profile from "./pages/Home/profile";
 import Chat from "./pages/Home/chat/chat.js";
 
 function App() {
