--- src/pages/Home/chat/chatStyles.css
@@ -1,118 +1,0 @@
-.chat_container {
-  border: 1px solid;
-  padding: 5px;
-  border-radius: 5px;
-  width: 100%;
-  background-color: #f5f5f5;
-  position: relative;
-}
-
-.chat_window {
-  border: 1px solid;
-  height: 448px;
-  margin-bottom: 5px;
-  overflow: scroll;
-  background-color: azure;
-  border-radius: 3px;
-  padding: 10px 0 0 20px;
-}
-
-#chat_form {
-  position: relative;
-  height: 30px;
-}
-.chat_send_btn {
-  background-color: #009f81;
-  height: 32px;
-  width: 36px;
-  border: none;
-  cursor: pointer;
-  position: absolute;
-  margin-left: 10px;
-  color: white;
-  font-size: 8px;
-  border-radius: 0 0 10px;
-}
-
-.chat_input {
-  display: table-cell;
-  height: 30px;
-  border: initial;
-  width: 90%;
-  border-radius: 8px;
-  background-color: white;
-  outline: none;
-}
-
-.chat_text_area {
-  margin-top: 10px;
-
-  position: relative;
-  max-width: 40%;
-  padding: 10px 10px;
-  margin-bottom: 10px;
-  background-color: #009f83;
-  display: table;
-  border-radius: 20px;
-  color: #ccf7f5;
-  font-size: 11px;
-  vertical-align: middle;
-  word-wrap: break-word;
-  margin-left: 3px;
-}
-
-.chat_text_container {
-  padding-bottom: 75px;
-}
-
-.chat_remove_message {
-  font-size: 10px;
-  position: absolute;
-  left: -21px;
-  cursor: pointer;
-  height: 20px;
-  width: 20px;
-  border-radius: 50%;
-  top: 20%;
-  padding: 0;
-}
-
-.chat_remove_gift {
-  font-size: 10px;
-  position: relative;
-  cursor: pointer;
-  height: 20px;
-  width: 20px;
-  border-radius: 50%;
-  top: 20%;
-  padding: 0;
-}
-
-.chat_remove_gift:hover {
-  background:#d8222c;
-}
-
-.chat_remove_message:hover {
-  background:#d8222c;
-}
-
-
-.chat_fake_textarea {
-  margin-top: 10px;
-  position: relative;
-  max-width: 40%;
-  padding: 10px 10px;
-  margin-bottom: 10px;
-  background-color: purple;
-
-  border-radius: 20px;
-  color: #ccf7f5;
-  font-size: 11px;
-  vertical-align: middle;
-  word-wrap: break-word;
-  margin-left: 3px;
-  display: flex;
-  justify-content: flex-end;
-
-
-}
