+++ src/pages/Home/chat/components/chat-form/index.js
@@ -1,4 +1,5 @@
 import React from "react";
+import "./styles.css";
 
 export function ChatForm({ sendMessage, inputRef, textValue, setTextValue }) {
   return (
