--- src/pages/Home/chat/emojiStyles.css
@@ -1,19 +1,0 @@
-.emoji_drawer {
-  background-color: cyan;
-    margin-top: -100px;
-    position: fixed;
-    border: 1px solid;
-    width: 5em;
-    right: 0;
-
-
-
-
-}
-
-.emoji_picker {
-  float: right;
-
-
-  transform: translateX(-50px);
-}
