+++ src/pages/Home/chat/styles.css
@@ -1,0 +1,8 @@
+.chat_container {
+  border: 1px solid;
+  padding: 5px;
+  border-radius: 5px;
+  width: 100%;
+  background-color: #f5f5f5;
+  position: relative;
+}
