+++ src/pages/Home/header.js
@@ -1,7 +1,5 @@
-import React from 'react';
-import './homeStyles.css'
-
-
+import React from "react";
+import "./styles.css";
 
 export default function Header() {
   return (
@@ -14,4 +12,3 @@ export default function Header() {
     </div>
   );
 }
-
